Fixing some signed/unsigned issues (exposed by gcc -W).
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@3826 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/devices/dev_asc.c
@@ -25,7 +25,7 @@
  *  SUCH DAMAGE.
  *
  *
- *  $Id: dev_asc.c,v 1.77 2005-12-11 16:45:58 debug Exp $
+ *  $Id: dev_asc.c,v 1.78 2005-12-26 14:22:32 debug Exp $
  *
  *  'asc' SCSI controller for some DECstation/DECsystem models and PICA-61.
  *
@@ -312,8 +312,8 @@ fatal("TODO..............\n");
 				fatal("no incoming DMA data?\n");
 				res = 0;
 			} else {
-				int len = d->xferp->data_in_len;
-				int len2 = d->reg_wo[NCR_TCL] +
+				size_t len = d->xferp->data_in_len;
+				size_t len2 = d->reg_wo[NCR_TCL] +
 				    d->reg_wo[NCR_TCM] * 256;
 				if (len2 == 0)
 					len2 = 65536;
